+++ src/presentation/erros/index.ts
@@ -1,0 +1,3 @@
+export * from "./invalid-param-error";
+export * from "./missing-param-error";
+export * from "./serverError";
